Review feedback applied
